made the components and xmacro def
