+++ BessSimEngine/include/component_definition.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "bess_api.h"
-#include "component_types.h"
+#include "component_types/component_types.h"
 #include "properties.h"
 #include "types.h"
 #include <string>
