+++ BessSimEngine/include/component_types/component_types.def
@@ -1,0 +1,26 @@
+
+COMPONENT(INPUT, )
+COMPONENT(OUTPUT, )
+COMPONENT(AND, )
+COMPONENT(OR, )
+COMPONENT(NOT, )
+COMPONENT(NOR, )
+COMPONENT(NAND, )
+COMPONENT(XOR, )
+COMPONENT(XNOR, )
+COMPONENT(FLIP_FLOP_JK, = 20)
+COMPONENT(FLIP_FLOP_SR, )
+COMPONENT(FLIP_FLOP_D, )
+COMPONENT(FLIP_FLOP_T, )
+COMPONENT(FULL_ADDER, )
+COMPONENT(HALF_ADDER, )
+COMPONENT(MULTIPLEXER_4_1, )
+COMPONENT(DECODER_2_4, )
+COMPONENT(DEMUX_1_4, )
+COMPONENT(ENCODER_4_2, )
+COMPONENT(HALF_SUBTRACTOR, )
+COMPONENT(MULTIPLEXER_2_1, )
+COMPONENT(COMPARATOR_1_BIT, )
+COMPONENT(PRIORITY_ENCODER_4_2, )
+COMPONENT(FULL_SUBTRACTOR, )
+COMPONENT(COMPARATOR_2_BIT, )
