+++ BessSimEngine/include/entt_components.h
@@ -2,7 +2,7 @@
 
 #include "bess_api.h"
 #include "bess_uuid.h"
-#include "component_types.h"
+#include "component_types/component_types.h"
 #include "types.h"
 #include <entt/entt.hpp>
 #include <utility>
